+++ src/components/FormControl/FormControl.stories.tsx
@@ -1,7 +1,10 @@
+import { generateTheme } from "@/lib/ai";
 import FormControl from "../FormControl/FormControl";
 import Input from "../Input/Input";
 
 import type { Meta, StoryObj } from "@storybook/react";
+import { useState } from "react";
+import Button from "../Button/Button";
 
 const meta: Meta<typeof FormControl> = {
   title: "Forms/FormControl",
@@ -33,5 +36,36 @@ export const WithTextarea: Story = {
   args: {
     title: "Description // wip",
     children: <textarea placeholder="Enter a detailed description" rows={4} />,
+  },
+};
+
+export const withAIGeneratedTheme: Story = {
+  render: () => {
+    const [value, setValue] = useState("");
+
+    const onClick = () => {
+      console.log(value);
+      generateTheme({
+        apiKey: import.meta.env.STORYBOOK_OPENAI_API_KEY!,
+        prompt: value,
+        options: {
+          dangerouslyAllowBrowser: true,
+        },
+      });
+    };
+    return (
+      <>
+        <FormControl title="With AI Generated Theme">
+          <Input
+            placeholder="fairy tale"
+            value={value}
+            onChange={(e) => setValue(e.target.value)}
+          />
+        </FormControl>
+        <Button variant="filled" isDisabled={false} onPress={onClick}>
+          Generate
+        </Button>
+      </>
+    );
   },
 };