+++ src/components/index.ts
@@ -20,7 +20,9 @@ import Container from "@/components/Container/Container";
 import Flex from "@/components/Flex/Flex";
 import FormControl from "@/components/FormControl/FormControl";
 import Heading from "@/components/Heading/Heading";
+import Image from "@/components/Image/Image";
 import Input from "@/components/Input/Input";
+import Paragraph from "@/components/Paragraph/Paragraph";
 import Radio from "@/components/Radio/Radio";
 import RadioGroup from "@/components/RadioGroup/RadioGroup";
 import Slider from "@/components/Slider/Slider";
@@ -45,8 +47,10 @@ export {
   Flex,
   FormControl,
   Heading,
+  Image,
   Input,
   Link,
+  Paragraph,
   Radio,
   RadioGroup,
   Slider,