+++ src/lib/ai.ts
@@ -4,7 +4,7 @@ import { generateColorsMap } from "./color";
 type Props = {
   apiKey: string;
   prompt: string;
-  options: Omit<ClientOptions, "apiKey">;
+  options?: Omit<ClientOptions, "apiKey">;
 };
 
 type Response = {
@@ -34,6 +34,8 @@ const generateTheme = async ({
           The customer gives you a word or tastes about the ambience of the site, return the best value for all variables.
           Consider contrast of text and back.
           The values should follow the format shown how.
+          Size should be return in rem.
+          Font family should be return serif or sans-serif.
           
           # Variables
           
@@ -41,6 +43,8 @@ const generateTheme = async ({
           --color-secondary
           --color-white
           --color-black
+          --radius-size
+          --font-family
           
           # Response rule
           
@@ -91,8 +95,8 @@ const generateTheme = async ({
       if (v.name === "--color-primary" || v.name === "--color-secondary") {
         const colors = generateColorsMap(v.color).colors;
         colors.forEach((c, i) => {
-          console.log(`--color-primary-${i}`, c.hex());
-          r.style.setProperty(`--color-primary-${i}`, c.hex());
+          console.log(`${v.name}-${i}`, c.hex());
+          r.style.setProperty(`${v.name}-${i}`, c.hex());
         });
       }
     });